tests/api_replicate.cc: Always use off_t to refer to the size of a file: fixes warnings about comparting a signed value to an unsigned value.
git-svn-id: file:///home/xapian-svn/xapian/trunk@14176 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/tests/api_replicate.cc
@@ -94,8 +94,8 @@ static void do_write(int fd, const char * p, size_t n)
 }
 
 // Make a truncated copy of a file.
-static size_t
-truncated_copy(const string & srcpath, const string & destpath, size_t tocopy)
+static off_t
+truncated_copy(const string & srcpath, const string & destpath, off_t tocopy)
 {
     int fdin = open(srcpath.c_str(), O_RDONLY);
     if (fdin == -1) {
@@ -405,8 +405,8 @@ replicate_with_brokenness(Xapian::DatabaseMaster & master,
     // Try applying truncated changesets of various different lengths.
     string brokenchangesetpath = tempdir + "/changeset_broken";
     off_t filesize = file_size(changesetpath);
-    size_t len = 10;
-    size_t copylen;
+    off_t len = 10;
+    off_t copylen;
     while (true) {
 	copylen = truncated_copy(changesetpath, brokenchangesetpath, len);
 	TEST_EQUAL(copylen, len);
